fix(openflare): serialize access log snowflake IDs as strings
- Return AccessLogView.id as string to avoid JS Number precision loss
- Store OpenFlareAccessLog IDs as uint64 with json id,string
- Update frontend AccessLogItem.id type to string

+++ internal/model/openflare_access_log.go
@@ -476,7 +476,7 @@ func openFlareAccessLogStatusCodeToInt32(code int) int32 {
 	}
 }
 
-func openFlareAccessLogUintToInt64(value uint) int64 {
+func openFlareAccessLogUintToInt64(value uint64) int64 {
 	if value > math.MaxInt64 {
 		return math.MaxInt64
 	}
